requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
colorgram.py==1.2.0
colorlog==6.7.0
colormap==1.0.4
contourpy==1.0.6
convcolors==2.2.0
cycler==0.11.0
dnspython==2.2.1
easydev==0.12.0
email-validator==1.3.0
extcolors==1.0.0
fastapi==0.89.1
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.3
numpy==1.24.1
opencv-python==4.7.0.68
orjson==3.8.5
packaging==23.0
pandas==1.5.2
pexpect==4.8.0
Pillow==9.4.0
ptyprocess==0.7.0
pybase64==1.2.3
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
